backend/requirements.txt
anyio==3.6.2
click==8.1.3
et-xmlfile==1.1.0
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
numpy==1.24.2
openpyxl==3.1.1
pandas==1.5.3
psycopg2==2.9.5
pydantic==1.10.5
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
